move format testutils to internal package
