+++ syft/formats/internal/testutils/utils.go
